ログイン機能を追加 authファイルを追加

+++ app/Http/Controllers/sns/DeleteController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\sns;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
+use App\Services\TweetService;
 
 class DeleteController extends Controller
 {
@@ -14,9 +15,13 @@ class DeleteController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request,TweetService $tweetService)
     {
         $tweetId=(int) $request->route('tweetId');
+        if(!$tweetService->checkUserId($request->user()->id,$tweetId))
+        {
+            return "他のユーザーの投稿は操作することはできません";
+        }
         $tweet=Tweet::where('id',$tweetId)->first();
         $tweet->delete();
